# crates/server/openapi.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: Warg Registry API
  description: |
    [warg](https://warg.io/) is an open source protocol for WebAssembly component registries.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0

# The Warg APIs currently do not use an authentication scheme
# Publishing requires that entries are signed with an acceptable key
...
servers:
  - url: http://localhost:8090/v1
    description: Local development server
paths:
...
  /proof/consistency:
    post:
      summary: Prove registry checkpoint consistency.
      operationId: proveConsistency
      security: []
      tags:
        - proof
      description: |
        Proves the consistency of the registry between two specified checkpoints.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProveConsistencyRequest"
      responses:
        "200":
          description: The consistency proof was generated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProveConsistencyResponse"
        "404":
          description: A requested entity was not found.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - status
                  - type
                  - id
                properties:
                  status:
                    type: integer
                    description: The HTTP status code for the error.
                    example: 404
                  type:
                    type: string
                    description: The type of entity that was not found.
                    enum: [logLength]
                    example: logLength
                  id:
                    type: integer
                    description: The identifier of the entity that was not found. 
        "422":
          description: The proof bundle could not be generated.
          content:
            application/json:
              schema:
                oneOf:
                  - "$ref": "#/components/schemas/BundleFailureError"
                discriminator:
                  propertyName: reason
                  mapping:
                    failure: "#/components/schemas/BundleFailureError"
        default:
          description: An error occurred when processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /proof/inclusion:
    post:
      summary: Prove log leaf inclusion.
      operationId: proveInclusion
      security: []
      tags:
        - proof
      description: |
        Proves that the given log leafs are present in the given registry checkpoint.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProveInclusionRequest"
      responses:
        "200":
          description: The inclusion proof was generated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProveInclusionResponse"
        "404":
          description: A requested entity was not found.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - status
                  - type
                  - id
                properties:
                  status:
                    type: integer
                    description: The HTTP status code for the error.
                    example: 404
                  type:
                    type: string
                    description: The type of entity that was not found.
                    enum: [logLength, leaf]
                    example: logLength
                  id:
                    type: integer
                    description: The identifier of the entity that was not found.
        "422":
          description: The proof bundle could not be generated.
          content:
            application/json:
              schema:
                oneOf:
                  - "$ref": "#/components/schemas/PackageNotIncludedError"
                  - "$ref": "#/components/schemas/IncorrectProofError"
                  - "$ref": "#/components/schemas/BundleFailureError"
                discriminator:
                  propertyName: reason
                  mapping:
                    packageNotIncluded: "#/components/schemas/PackageNotIncludedError"
                    incorrectProof: "#/components/schemas/IncorrectProofError"
                    failure: "#/components/schemas/BundleFailureError"
        default:
          description: An error occurred when processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Error:
      type: object
      description: A generic error response.
      additionalProperties: false
      required:
        - status
        - message
      properties:
        status:
          type: integer
          description: The HTTP status code for the error.
          example: 406
        message:
          type: string
          description: The error message.
          example: the server cannot produce an acceptable response
    AnyHash:
      type: string
      description: Represents a supported hash.
      example: sha256:b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944c
      pattern: ^[a-z0-9-]+:[a-f0-9]+$
...
    ProveConsistencyRequestParams:
      type: object
      description: Consistency proof request params.
      additionalProperties: false
      required:
        - from
        - to
      properties:
        from:
          type: integer
          description: The starting log length.
          minimum: 1
          example: 42
        to:
          type: integer
          description: The ending log length.
          minimum: 1
          example: 42
    ProveConsistencyRequest:
      type: object
      description: A request to prove the consistency of the registry.
      additionalProperties: false
      allOf:
        - $ref: "#/components/schemas/ProveConsistencyRequestParams"
        - type: object
          additionalProperties: false
          federated:
            type: object
            description: For federated packages, the params for consistency proof request.
            additionalProperties:
              type: object
              $ref: "#/components/schemas/ProveConsistencyRequestParams"
    ProveConsistencyResponse:
      type: object
      description: A response containing the consistency proof bundle.
      additionalProperties: false
      required:
        - proof
      properties:
        proof:
          type: string
          description: The consistency proof bundle.
          format: byte
          example: "ZXhhbXBsZQ=="
        federated:
          type: object
          description: For federated packages, the consistency proof bundles.
          additionalProperties:
            type: string
            description: The consistency proof bundle.
            format: byte
            example: "ZXhhbXBsZQ=="
    ProveInclusionRequestParams:
      type: object
      description: Inclusion proof request params.
      additionalProperties: false
      required:
        - logLength
        - leafs
      properties:
        logLength:
          type: integer
          description: The checkpoint log length to prove the inclusion for.
        leafs:
          type: array
          maxItems: 1000
          description: The log leaf registry log index to prove the inclusion for.
          items:
            type: integer
    ProveInclusionRequest:
      type: object
      description: A request to prove the inclusion of log leafs in a checkpoint.
      additionalProperties: false
      allOf:
        - $ref: "#/components/schemas/ProveInclusionRequestParams"
        - type: object
          additionalProperties: false
          federated:
            type: object
            description: For federated packages, the params for inclusion proof request.
            additionalProperties:
              type: object
              $ref: "#/components/schemas/ProveInclusionRequestParams"
    ProveInclusionProofBundles:
      type: object
      description: The inclusion proof bundle.
      additionalProperties: false
      required:
        - log
        - map
      properties:
        log:
          type: string
          description: The log inclusion proof bundle.
          format: byte
          example: "ZXhhbXBsZQ=="
        map:
          type: string
          description: The map inclusion proof bundle.
          format: byte
          example: "ZXhhbXBsZQ=="
    ProveInclusionResponse:
      type: object
      description: A response containing the inclusion proof bundle.
      additionalProperties: false
      allOf:
        - $ref: "#/components/schemas/ProveInclusionProofBundles"
        - type: object
          additionalProperties: false
          federated:
            type: object
            description: For federated packages, the inclusion proof bundles.
            additionalProperties:
              type: object
              $ref: "#/components/schemas/ProveInclusionProofBundles"
...
    PackageNotIncludedError:
      type: object
      additionalProperties: false
      required:
        - status
        - reason
        - logId
      properties:
        status:
          type: integer
          description: The HTTP status code for the error.
          example: 422
        reason:
          type: string
          description: The reason why the bundle could not be generated.
          enum: [packageNotIncluded]
          example: packageNotIncluded
        logId:
          "$ref": "#/components/schemas/AnyHash"
          description: The identifier of the log that was not included.
    IncorrectProofError:
      type: object
      additionalProperties: false
      required:
        - status
        - reason
        - root
        - found
      properties:
        status:
          type: integer
          description: The HTTP status code for the error.
          example: 422
        reason:
          type: string
          description: The reason why the bundle could not be generated.
          enum: [incorrectProof]
          example: incorrectProof
        root:
          "$ref": "#/components/schemas/AnyHash"
          description: The map or log root hash from the checkpoint that was provided.
        found:
          "$ref": "#/components/schemas/AnyHash"
          description: The hash that was found in the proof.
    BundleFailureError:
      type: object
      additionalProperties: false
      required:
        - status
        - reason
        - message
      properties:
        status:
          type: integer
          description: The HTTP status code for the error.
          example: 422
        reason:
          type: string
          description: The reason why the bundle could not be generated.
          enum: [failure]
          example: failure
        message:
          type: string
          description: The failure error message.
          example: bundle must contain proofs for the same root
...